BUILDBOT: Correct cleanup script to work properly.
This now omits the root directory (which I think triggered the removal
of the "aged" stable directory) and fixes the logic on the
subdirectories which were allowing accumulation of builds in master...

+++ wwwclean.sh
@@ -14,16 +14,18 @@ SNAPSHOT_DIR=/var/www/snapshots/
 # Iterate across snapshot directories, removing older builds..
 # BUT skipping removal if they are the only remaining build.
 for snapsubdir in `find ${SNAPSHOT_DIR} -type d`; do
-	# If number of builds determined by native build name
-	# is greater than 2, remove older ones ie. Just current and symlink 
-	# will be left
-	NATIVE_BUILD="debian-x86-`basename ${snapsubdir}`"
-	NUM_OF_BUILDS=`find ${NATIVE_BUILD}-* | wc -l`
-	if [ "${NUM_OF_BUILDS}" -gt 2 ]; then
-		# Remove any files older than KEEP_N_DAYS
-		find ${snapsubdir} -type f -mtime +${KEEP_N_DAYS} -delete
-		# Remove any symbolic links older than KEEP_N_DAYS
-		find ${snapsubdir} -type l -mtime +${KEEP_N_DAYS} -delete
+	if [ ${snapsubdir} != ${SNAPSHOT_DIR} ]; then
+		# If number of builds determined by native build name
+		# is greater than 2, remove older ones ie. Just current and symlink 
+		# will be left
+		NATIVE_BUILD="debian-x86-`basename ${snapsubdir}`"
+		NUM_OF_BUILDS=`find ${snapsubdir}/${NATIVE_BUILD}-* | wc -l`
+		if [ "${NUM_OF_BUILDS}" -gt 2 ]; then
+			# Remove any files older than KEEP_N_DAYS
+			find ${snapsubdir} -type f -mtime +${KEEP_N_DAYS} -delete
+			# Remove any symbolic links older than KEEP_N_DAYS
+			find ${snapsubdir} -type l -mtime +${KEEP_N_DAYS} -delete
+		fi
 	fi
 done
